fix: fix code and add notifications
